chefsalot-server/chefsalotserver/requirements.txt
Django==2.0.6
django-cors-headers==2.3.0
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.4
